+++ azure-pipelines.yml
@@ -10,10 +10,10 @@ jobs:
     matrix:
       SQL-2019:
         Target_SQL: 'HGS-2k19-01'
-        Ex_Groups: 'xSQLv15,MSI,clientCertAuth'
+        Ex_Groups: 'xSQLv15,MSI,clientCertAuth,Fedauth'
       SQL-2012:
         Target_SQL: 'SQL-2K12-SP3-1'
-        Ex_Groups: 'xSQLv12,MSI'
+        Ex_Groups: 'xSQLv12,MSI,Fedauth'
     maxParallel: 2
   steps:
   - powershell: |
@@ -29,10 +29,6 @@ jobs:
       Get-Content .\JavaKeyStoreBase.txt | Set-Content -Encoding utf8 JavaKeyStore.txt
       Remove-Item –path .\JavaKeyStoreBase.txt
     displayName: 'PowerShell Script'
-  - task: DownloadSecureFile@1
-    displayName: 'Download config.properties'
-    inputs:
-      secureFile: config.properties
   - task: DownloadSecureFile@1
     name: pkcs12_truststore
     displayName: 'Download PKCS12 truststore file'
