+++ src/test/java/com/microsoft/sqlserver/jdbc/AlwaysEncrypted/AESetup.java
@@ -273,6 +273,35 @@ public class AESetup extends AbstractTest {
     }
 
     /**
+     * Read the alias from file which is created during creating jks If the jks and alias name in JavaKeyStore.txt does
+     * not exists, will not run!
+     * 
+     * @param inputFile
+     * @param lookupValue
+     * @throws IOException
+     */
+    protected static void readFromFile(String inputFile, String lookupValue) throws IOException {
+        String filePath = TestUtils.getCurrentClassPath();
+        try {
+            File f = new File(filePath + inputFile);
+            try (BufferedReader buffer = new BufferedReader(new FileReader(f))) {
+                String readLine = "";
+                String[] linecontents;
+
+                while ((readLine = buffer.readLine()) != null) {
+                    if (readLine.trim().contains(lookupValue)) {
+                        linecontents = readLine.split(" ");
+                        javaKeyAliases = linecontents[2];
+                        break;
+                    }
+                }
+            }
+        } catch (IOException e) {
+            fail(e.getMessage());
+        }
+    }
+
+    /**
      * Create AE test tables
      * 
      * @param tableName
