+++ src/test/java/com/microsoft/sqlserver/jdbc/fedauth/ConnectionEncryptionTest.java
@@ -15,6 +15,7 @@ import java.sql.Statement;
 
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
 import org.junit.jupiter.api.Test;
 import org.junit.platform.runner.JUnitPlatform;
 import org.junit.runner.RunWith;
@@ -23,10 +24,10 @@ import com.microsoft.sqlserver.jdbc.RandomUtil;
 import com.microsoft.sqlserver.jdbc.SQLServerException;
 import com.microsoft.sqlserver.jdbc.TestResource;
 import com.microsoft.sqlserver.jdbc.TestUtils;
-import com.microsoft.sqlserver.testframework.AbstractTest;
-
+import com.microsoft.sqlserver.testframework.Constants;
 
 @RunWith(JUnitPlatform.class)
+@Tag(Constants.Fedauth)
 public class ConnectionEncryptionTest extends FedauthCommon {
 
     static String charTable = RandomUtil.getIdentifier("charTableFedAuth");
