+++ src/test/java/com/microsoft/sqlserver/jdbc/fedauth/FedauthCommon.java
@@ -6,13 +6,14 @@ import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 
 import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.Tag;
 
 import com.microsoft.aad.adal4j.AuthenticationContext;
 import com.microsoft.aad.adal4j.AuthenticationResult;
 import com.microsoft.sqlserver.testframework.Constants;
 import com.microsoft.sqlserver.testframework.AbstractTest;
 
-
+@Tag(Constants.Fedauth)
 public class FedauthCommon extends AbstractTest {
 
     public static String azureServer = null;
@@ -22,7 +23,6 @@ public class FedauthCommon extends AbstractTest {
     public static String azureGroupUserName = null;
 
     public static boolean enableADIntegrated = false;
-
     public static String spn = null;
     public static String stsurl = null;
     public static String fedauthClientId = null;
@@ -49,7 +49,7 @@ public class FedauthCommon extends AbstractTest {
 
         enableADIntegrated = !System.getProperty("os.name").startsWith(
                 "Windows") ? false
-                           : getConfiguredProperty("enableADIntegrated").equalsIgnoreCase("true") ? true : false;
+                           : getConfiguredProperty("enableADIntegrated", "").equalsIgnoreCase("true") ? true : false;
 
         hostNameInCertificate = getConfiguredProperty("hostNameInCertificate");
