+++ src/test/java/com/microsoft/sqlserver/testframework/Constants.java
@@ -25,6 +25,7 @@ public final class Constants {
      * NTLM  - - - - - - - For NTLM tests
      * reqExternalSetup  - For tests requiring external setup
      * clientCertAuth  - - For tests requiring client certificate authentication setup
+     * Fedauth - - - - - - For Fedauth tests
      * </pre>
      */
     public static final String xJDBC42 = "xJDBC42";
@@ -39,6 +40,7 @@ public final class Constants {
     public static final String MSI = "MSI";
     public static final String reqExternalSetup = "reqExternalSetup";
     public static final String clientCertAuth = "clientCertAuth";
+    public static final String Fedauth = "Fedauth";
 
     public static final ThreadLocalRandom RANDOM = ThreadLocalRandom.current();
     public static final Logger LOGGER = Logger.getLogger("AbstractTest");
